return to int amount

+++ herfaty/lib/MangeProduct.dart/EditProduct.dart
@@ -470,9 +470,8 @@ class _EditProduc extends State<EditProduct> {
                     validator: (value) {
                       if (value == null || value.isEmpty)
                         return 'أدخل كمية المنتج';
-                      if (double.parse(value) <= 0)
-                        return "أدخل رقم أكبر من صفر";
-                      if (double.parse(value) > 15)
+                      if (int.parse(value) <= 0) return "أدخل رقم أكبر من صفر";
+                      if (int.parse(value) > 15)
                         return "أدخل رقم أصغر من أو يساوي 15";
                       else
                         return null;
